+++ AutomatedLab/AutomatedLabVirtualMachines.psm1
@@ -2419,39 +2419,23 @@ function Copy-LabALCommon
         [string[]]
         $ComputerName
     )
+
+    $machines = Get-LabVm -ComputerName $ComputerName | Where-Object -FilterScript {-not $_.InternalNotes.ALCommonCopied}
+
+    if (-not $machines) { return }
     
-    $childPath = foreach ($vm in $ComputerName)
+    $module = Get-Module -Name AutomatedLab.Common -ListAvailable | Sort-Object -Property Version | Select-Object -Last 1
+    Send-ModuleToPSSession -Module $module -Session (New-LabPSSession -ComputerName $machines) -Scope AllUsers -IncludeDependencies
+
+    foreach ($machine in $machines)
     {
-        Invoke-LabCommand -ScriptBlock {
-            if ($PSEdition -eq 'Core')
-            {
-                'core'
-            } else
-            {
-                'full'
-            }
-        } -ComputerName $vm -NoDisplay -IgnoreAzureLabSources -PassThru |
-        Add-Member -MemberType NoteProperty -Name ComputerName -Value $vm -Force -PassThru
+        $shouldExport = $true
+        $machine.InternalNotes.Add('ALCommonCopied', $true)
     }
 
-    $coreChild = @($childPath) -eq 'core'
-    $fullChild = @($childPath) -eq 'full'
-    $libLocation = Split-Path -Parent -Path (Split-Path -Path ([AutomatedLab.Common.Win32Exception]).Assembly.Location -Parent)
-
-    if ($coreChild -and @(Invoke-LabCommand -ScriptBlock{
-                Get-Item -Path '/ALLibraries/core/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-    } -ComputerName $coreChild.ComputerName -IgnoreAzureLabSources -NoDisplay -PassThru).Count -ne $coreChild.Count)
+    if ($shouldExport)
     {
-        $coreLibraryFolder = Join-Path -Path $libLocation -ChildPath $coreChild[0]
-        Copy-LabFileItem -Path $coreLibraryFolder -ComputerName $coreChild.ComputerName -DestinationFolderPath '/ALLibraries' -UseAzureLabSourcesOnAzureVm $false
-    }
-
-    if ($fullChild -and @(Invoke-LabCommand -ScriptBlock {
-                Get-Item -Path '/ALLibraries/full/AutomatedLab.Common.dll' -ErrorAction SilentlyContinue
-    } -ComputerName $fullChild.ComputerName -IgnoreAzureLabSources -NoDisplay -PassThru).Count -ne $fullChild.Count)
-    {
-        $fullLibraryFolder = Join-Path -Path $libLocation -ChildPath $fullChild[0]
-        Copy-LabFileItem -Path $fullLibraryFolder -ComputerName $fullChild.ComputerName -DestinationFolderPath '/ALLibraries' -UseAzureLabSourcesOnAzureVm $false
+        Export-Lab
     }
 }
 #endregion Copy-LabALCommon